+++ src/Engine/Graphics/Collisions.cpp
@@ -5,7 +5,7 @@
 #include <utility>
 
 #include "Engine/Events/Processor.h"
-#include "Engine/Graphics/DecorationList.h"
+#include "Engine/Objects/DecorationList.h"
 #include "Engine/Graphics/Level/Decoration.h"
 #include "Engine/Graphics/Outdoor.h"
 #include "Engine/Graphics/Indoor.h"
