+++ src/Engine/Graphics/Sprites.cpp
@@ -6,7 +6,7 @@
 #include <string>
 
 #include "Engine/OurMath.h"
-#include "Engine/Graphics/DecorationList.h"
+#include "Engine/Objects/DecorationList.h"
 #include "Engine/Graphics/PaletteManager.h"
 #include "Engine/Graphics/Image.h"
 #include "Engine/LodSpriteCache.h"
