+++ src/Engine/Objects/CMakeLists.txt
@@ -4,6 +4,7 @@ set(ENGINE_OBJECTS_SOURCES
         Actor.cpp
         Chest.cpp
         CombinedSkillValue.cpp
+        DecorationList.cpp
         ItemEnumFunctions.cpp
         Items.cpp
         MonsterEnumFunctions.cpp
@@ -20,6 +21,7 @@ set(ENGINE_OBJECTS_HEADERS
         Chest.h
         ChestEnums.h
         CombinedSkillValue.h
+        DecorationList.h
         DecorationEnums.h
         ItemEnchantment.h
         ItemEnums.h
